pyproject.toml
[tool.poetry]
name = "mlsystemdesign-mlserveryolov8"
version = "0.1.0"
description = ""
authors = ["Artemiy Mazaew <mr8bit@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.10.13"
mlserver = "^1.5.0"
ultralytics = "^8.2.12"
pillow = "^10.3.0"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
